backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
pydantic
chromadb
sentence-transformers
pypdf
python-docx
Pillow
